implement offsets-specific asvs; flake8 cleanup
